rgw: move svc_cls -> rados/svc_cls
