+++ src/rgw/CMakeLists.txt
@@ -25,7 +25,6 @@ set(librgw_common_srcs
   services/svc_bucket.cc
   services/svc_bucket_sobj.cc
   services/svc_bucket_sync_sobj.cc
-  services/svc_cls.cc
   services/svc_config_key_rados.cc
   services/svc_mdlog.cc
   services/svc_meta.cc
@@ -45,6 +44,7 @@ set(librgw_common_srcs
   services/svc_zone.cc
   services/rados/svc_bi_rados.cc
   services/rados/svc_bilog_rados.cc
+  services/rados/svc_cls.cc
   services/rados/svc_rados.cc
   services/svc_zone_utils.cc
   rgw_service.cc
