+++ src/rgw/rgw_metadata.cc
@@ -19,7 +19,7 @@
 #include "services/svc_meta.h"
 #include "services/svc_meta_be.h"
 #include "services/svc_meta_be_sobj.h"
-#include "services/svc_cls.h"
+#include "services/rados/svc_cls.h"
 
 #include "include/ceph_assert.h"
