+++ src/rgw/rgw_sync.cc
@@ -29,7 +29,7 @@
 #include "services/svc_zone.h"
 #include "services/svc_mdlog.h"
 #include "services/svc_meta.h"
-#include "services/svc_cls.h"
+#include "services/rados/svc_cls.h"
 
 #include <boost/asio/yield.hpp>
