+++ src/rgw/services/rados/svc_cls.cc
