+++ src/rgw/services/rados/svc_cls.h
